requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
pylint==2.5.3
python-dotenv==0.14.0
requests==2.18.4
uWSGI==2.0.19.1
psycopg2-binary
bcrypt==3.1.7
PyJWT